refactor: Tests & overflow
