+++ cypress/integration/smoke.js
@@ -1,0 +1,7 @@
+import { getFirstFrame } from '../support/utils';
+
+describe('Smoke', () => {
+  it('frames are interactive', () => {
+    getFirstFrame().click('center');
+  });
+});
