+++ src/Playroom/Toolbar/Toolbar.less
@@ -3,10 +3,11 @@
 .root {
   position: relative;
   height: 100%;
-  min-width: @toolbar-closed-size;
+  min-width: @toolbar-closed-size + @toolbar-border-thickness;
   display: flex;
   flex-direction: row-reverse;
   color: @toolbar-foregound;
+  overflow: hidden;
 
   &.isOpen {
     width: 100vw;
